requirements.txt
# Core LangChain dependencies
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1

# Model providers
langchain-mistralai>=0.0.1
langchain-openai>=0.0.5
langchain-huggingface>=0.0.1

# Document processing
pypdf>=4.0.0
pdfplumber>=0.10.0

# Embeddings and ML
sentence-transformers>=2.2.0
torch>=2.0.0
transformers>=4.30.0

# Environment and utilities
python-dotenv>=1.0.0
numpy>=1.24.0

# Optional: for better PDF processing
pymupdf>=1.23.0
